refactor: tide up code a bit
