+++ next-docs/pages/components/textInputGroup.tsx
@@ -3,6 +3,7 @@ import Preview from '../../components/codePreview/Preview';
 import PropsTable from '../../components/PropsTable';
 import AccessValues from '../../public/examples/textInputGroup/AccessValues';
 import Combined from '../../public/examples/textInputGroup/Combined';
+import CombinedVertical from "../../public/examples/textInputGroup/CombinedVertical";
 import Default from '../../public/examples/textInputGroup/Default';
 import EventListenerSelect from '../../public/examples/textInputGroup/EventListenerSelect';
 import MultiSelect from '../../public/examples/textInputGroup/MultiSelect';
@@ -50,6 +51,14 @@ const PageTextInputGroup = () => {
           title="Combined"
           preview={<Combined />}
           code={examples ? examples.Combined : 'Loading'}
+        />
+      </section>
+
+      <section className="mt-8">
+        <Preview
+          title="Combined vertical"
+          preview={<CombinedVertical />}
+          code={examples ? examples.CombinedVertical : 'Loading'}
         />
       </section>
