+++ next-docs/public/examples/textInputGroup/SingleSelectRtl.tsx
@@ -19,17 +19,19 @@ const Example = () => {
 
   const selectProps1 = {
     options,
+    type: 'select',
     placeholderSlot: 'Choose an option',
   };
   const selectProps2 = {
     options,
+    type: 'select',
     label: 'xLarge',
     placeholderSlot: 'Choose an option',
   };
   return (
     <div className="flex items-end flex-wrap pb-44 w-full justify-center ps-40 pe-40">
       <TextInputGroup
-      inputProps={{
+        inputProps={{
           input1: selectProps1,
           input2: selectProps2,
         }}
