+++ packages/select/src/private/types/CombinedInputProps.ts
@@ -1,0 +1,4 @@
+import {TextInputProps} from "@heathmont/moon-core";
+import {SelectProps} from "../../Select";
+
+export type CombinedPropType = Partial<SelectProps<any>> & Partial<TextInputProps> & { type: string };
