+++ packages/select/src/private/types/CombinedInputSize.ts
@@ -1,0 +1,1 @@
+export type TextInputGroupSize = "medium" | "large";
